requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.3            # via geoip2
asgiref==3.3.1            # via django
async-timeout==3.0.1      # via aiohttp
attrs==20.3.0             # via aiohttp
cbor2==5.2.0              # via webauthn
certifi==2020.12.5        # via requests
cffi==1.14.4              # via cryptography
chardet==3.0.4            # via aiohttp, requests
click==7.1.2              # via geoip2-tools
cryptography==3.3.1       # via pyopenssl, webauthn
django-bitfield==2.0.1    # via -r requirements.in
django-recaptcha==2.0.6   # via -r requirements.in
django==3.1.4             # via -r requirements.in, django-bitfield, django-recaptcha, djangorestframework, jsonfield
djangorestframework==3.12.2  # via -r requirements.in
future==0.18.2            # via webauthn
geoip2-tools==0.1.1       # via -r requirements.in
geoip2==4.1.0             # via -r requirements.in, geoip2-tools
idna==2.10                # via requests, yarl
jsonfield==3.1.0          # via -r requirements.in
maxminddb==2.0.3          # via geoip2
multidict==5.1.0          # via aiohttp, yarl
pillow==8.1.0             # via -r requirements.in
pycparser==2.20           # via cffi
pyopenssl==20.0.1         # via webauthn
pyotp==2.4.1              # via -r requirements.in
pytz==2020.4              # via django
qrcode==6.1               # via -r requirements.in
requests==2.25.1          # via geoip2, geoip2-tools
six==1.15.0               # via cryptography, django-bitfield, pyopenssl, qrcode, webauthn
sqlparse==0.4.1           # via django
typing-extensions==3.7.4.3  # via aiohttp
urllib3==1.26.2           # via geoip2, requests
webauthn==0.4.7           # via -r requirements.in
yarl==1.6.3               # via aiohttp
